requirements.txt
astroid==2.2.5
Click==7.0
colorama==0.4.1
cycler==0.10.0
dash==1.0.2
dash-core-components==1.0.0
dash-daq==0.1.0
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.0.2
Django==2.2.4
Flask==1.1.1
Flask-Compress==1.4.0
gunicorn==19.7.1
image==1.5.27
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.0
Pillow==6.1.0
plotly==4.0.0
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
webcolors==1.9.1
Werkzeug==0.15.5
wrapt==1.11.2
